OPENEUROPA-2221: Add event kernel test base class.
